Add support for manual numbering and copying instead of moving
Now when the user passes `--match` and a regex, if the regex includes a capture group, the program will try to parse that captured value as an integer and use that as the number for that file, padding appropriately of course.

No change in behavior if a capture group is not specified or if the captured data doesn't parse as an integer.

The `--commit` argument was replaced by `--cp` and `--mv`, so the user can specify if they want the file copied or moved.